Rifle Firing Blendings Fix
When the character is aiming without walking, the blending is great, but
when the character is aiming and the character is moving at the same
time, then the firing montage is playing and affecting the whole upper
body, and the character is not aiming at the give direction. I had to
create a custom layered blend from bone, and play the firing montage so
that only the lower hands are affected while the character is moving.
Now everything works fine.